feat: initial working solution, returns merged accounts of related email
while the first account is entered into the 'mergedAccounts' array by simply
copying the existing data, each subsequent account is checked against the
mergedAccount array to see if there are any existing 'mergedAccounts' that have
matching emails to the current account being processed. if so, the indexes for
those matches within the 'mergedAccounts' list are retrieved, and, after retrieving
the relevant information (existing emails and applications), they are removed
from the 'mergedAccounts' list. instead, they are incorporated into the current
account being processed.