refactor: corrects misunderstanding, goal is to return 'People'
according to the specs, the goal is not to return 'accounts', but
to merge 'Account(s)', creating a 'Person'. the change doesn't require
any extra effort, other than separating the type definitions of Account
and MergedAccount -> Person. renames all of the types accordingly, renames
some of the variables in order to reflect the change in logic

creates:
src/types/person.d.ts